chore: revert interop changes
