+++ interop/Dockerfile
@@ -1,19 +1,14 @@
-# Here because we want to fetch the node_modules within docker so that it's
-# installed on the same platform the test is run. Otherwise tools like `esbuild` will fail to run
-FROM node:22-bullseye-slim
+FROM node:20
 
 WORKDIR /app
-
-COPY package.json ./
+COPY package.json .
+COPY ./node_modules ./node_modules
 COPY ./packages ./packages
 COPY ./interop ./interop
+
+WORKDIR /app/interop
 
 # disable colored output and CLI animation from test runners
 ENV CI true
 
-RUN npm i
-RUN npm run build
-
-WORKDIR /app/interop
-
-ENTRYPOINT npm test -- -t node
+ENTRYPOINT [ "npm", "run", "test:interop:multidim", "--", "--build", "false", "--types", "false", "-t", "node" ]
