+++ interop/package.json
@@ -1,8 +1,45 @@
 {
-  "name": "@libp2p/transport-interop-libp2p-main",
-  "version": "1.0.0",
+  "name": "@libp2p/multidim-interop",
+  "version": "2.0.33",
+  "description": "Multidimensional interop tests",
+  "author": "Glen De Cauwsemaecker <glen@littlebearlabs.io> / @marcopolo",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p/tree/main/interop#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/libp2p/js-libp2p.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p/issues"
+  },
+  "publishConfig": {
+    "access": "public",
+    "provenance": true
+  },
   "type": "module",
-  "private": true,
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "test",
+    "dist",
+    ".aegir.js",
+    "BrowserDockerfile",
+    "chromium-version.json",
+    "Dockerfile",
+    "firefox-version.json",
+    "Makefile",
+    "node-version.json",
+    "relay.js",
+    "tsconfig.json",
+    "webkit-version.json",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
   "eslintConfig": {
     "extends": "ipfs",
     "parserOptions": {
@@ -11,17 +48,16 @@
     }
   },
   "scripts": {
-    "clean": "aegir clean",
-    "build": "aegir build --bundle false",
-    "test": "aegir test",
+    "start": "node index.js",
+    "build": "aegir build",
     "lint": "aegir lint",
-    "dep-check": "aegir dep-check"
+    "test:interop:multidim": "aegir test",
+    "test:transport:interop": "aegir test"
   },
-  "devDependencies": {
+  "dependencies": {
     "@chainsafe/libp2p-noise": "^15.0.0",
     "@chainsafe/libp2p-yamux": "^6.0.2",
     "@libp2p/circuit-relay-v2": "^1.0.24",
-    "@libp2p/interface": "^1.4.0",
     "@libp2p/identify": "^2.0.2",
     "@libp2p/mplex": "^10.0.24",
     "@libp2p/ping": "^1.0.19",
@@ -29,10 +65,10 @@
     "@libp2p/webrtc": "^4.0.33",
     "@libp2p/websockets": "^8.0.24",
     "@libp2p/webtransport": "^4.0.32",
+    "@multiformats/mafmt": "^12.1.6",
     "@multiformats/multiaddr": "^12.2.3",
     "aegir": "^42.2.11",
     "libp2p": "^1.6.0",
-    "p-event": "^6.0.1",
     "redis": "^4.6.14"
   },
   "browser": {
