+++ interop/tsconfig.json
@@ -9,9 +9,6 @@
   ],
   "references": [
     {
-      "path": "../packages/interface"
-    },
-    {
       "path": "../packages/libp2p"
     },
     {
